+++ packages/documentation/docs/components/html-grid/code.mdx
@@ -1,6 +1,7 @@
 ---
 doc-type: 'tab-item'
 ---
+
 import HtmlTablePlayground from '@site/docs/autogenerated/playground/html-table.mdx';
 import HtmlTableStripedPlayground from '@site/docs/autogenerated/playground/html-table-striped.mdx';
 
@@ -8,8 +9,8 @@ import HtmlTableStripedPlayground from '@site/docs/autogenerated/playground/html
 
 ## Basic
 
-<HtmlTablePlayground height="12rem" />
+<HtmlTablePlayground height="12rem" onlyFramework="html" />
 
 ## Striped
 
-<HtmlTableStripedPlayground height="12rem" />
+<HtmlTableStripedPlayground height="12rem" onlyFramework="html" />
