+++ packages/documentation/docs/legacy/basic-navigation/code.mdx
@@ -1,5 +1,6 @@
 ---
 doc-type: 'tab-item'
+unlisted: true
 ---
 import PropsApi from '@site/docs/autogenerated/api/ix-basic-navigation/api.mdx';
