+++ packages/documentation/docs/legacy/basic-navigation/index.mdx
@@ -1,5 +1,5 @@
 ---
-doc-type: "tabs"
+doc-type: 'tabs'
 description: 'Basic navigation combines essential infrastructure components to form the foundational layout structure of an application.'
 title: 'Basic navigation'
 deprecated:
@@ -7,4 +7,3 @@ deprecated:
   - /docs/components/application
 unlisted: true
 ---
-
