+++ packages/documentation/docs/legacy/map-navigation/code.mdx
@@ -1,5 +1,6 @@
 ---
 doc-type: 'tab-item'
+unlisted: true
 ---
 import PropsApi from '@site/docs/autogenerated/api/ix-map-navigation/api.mdx';
 import PropsOverlayApi from '@site/docs/autogenerated/api/ix-map-navigation-overlay/api.mdx';
