+++ packages/documentation/src/components/theme/DocTabsHeader/index.tsx
@@ -1,12 +1,12 @@
 /*
  * COPYRIGHT (c) Siemens AG 2018-2024 ALL RIGHTS RESERVED.
  */
+import type { PropSidebarItemLink } from '@docusaurus/plugin-content-docs';
 import { useHistory, useLocation } from '@docusaurus/router';
 import { useScrollPosition } from '@docusaurus/theme-common/internal';
 import { DeprecatedTag, RedirectTag } from '@site/src/components/UI/Tags';
-import useSearchParams from '@site/src/utils/hooks/useSearchParams';
 import clsx from 'clsx';
-import { Fragment, useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 import styles from './styles.module.css';
 
 function Tabs({ children }) {
@@ -28,26 +28,16 @@ function Tabs({ children }) {
 
 function Tab(props: { label: string; value: string }) {
   const location = useLocation();
-  const currentTab = useSearchParams('current-tabs');
   const history = useHistory();
 
-  const [isActive, setIsActive] = useState(false);
-
   const onNavigate = useCallback(() => {
-    const searchParams = new URLSearchParams(location.search);
-    searchParams.set('current-tabs', props.value);
-    location.search = searchParams.toString();
-    history.push(location);
-  }, [props.value, location]);
-
-  useEffect(() => {
-    setIsActive(props.value === currentTab);
-  }, [location, currentTab]);
+    history.push(props.value);
+  }, [props.value, history]);
 
   return (
     <div
       className={clsx(styles.Tab, {
-        [styles['Tab--active']]: isActive,
+        [styles['Tab--active']]: location.pathname === props.value,
       })}
       onClick={onNavigate}
     >
@@ -56,15 +46,14 @@ function Tab(props: { label: string; value: string }) {
   );
 }
 
-export default function HeroHeader(props: {
+export default function DocTabsHeader(props: {
   id: string;
   title: string;
   description: string;
-  tabs: string[];
+  tabs: PropSidebarItemLink[];
   frontMatter: any;
 }) {
   const { description, tabs, title, frontMatter, id } = props;
-  const noSingleTab = props.frontMatter.no_single_tab;
 
   return (
     <>
@@ -113,11 +102,7 @@ export default function HeroHeader(props: {
 
       <Tabs>
         {tabs.map((tab, index) => {
-          if (tabs.length > 0 && !noSingleTab) {
-            return <Tab value={tab} label={tab} key={tab} />;
-          }
-
-          return <Fragment key={index}></Fragment>;
+          return <Tab value={tab.href} label={tab.label} key={tab.docId} />;
         })}
       </Tabs>
     </>
