+++ packages/documentation/src/theme/DocItem/Layout/index.tsx
@@ -1,14 +1,15 @@
 import {
-  useDoc,
   useCurrentSidebarCategory,
+  useDoc,
   useDocById,
 } from '@docusaurus/plugin-content-docs/client';
 import { useHistory } from '@docusaurus/router';
 import { useWindowSize } from '@docusaurus/theme-common';
 
+import { PropSidebarItemLink } from '@docusaurus/plugin-content-docs';
 import type { WrapperProps } from '@docusaurus/types';
-import HeroHeader from '@site/src/components/theme/HeroHeader';
-import { useSubPageHook } from '@site/src/components/theme/QueryStringContent';
+import DocDefaultHeader from '@site/src/components/theme/DocDefaultHeader';
+import DocTabsHeader from '@site/src/components/theme/DocTabsHeader';
 import { useDocType } from '@site/src/utils/hooks/useDocType';
 import DocItemContent from '@theme-original/DocItem/Content';
 import DocItemFooter from '@theme-original/DocItem/Footer';
@@ -18,23 +19,15 @@ import type LayoutType from '@theme/DocItem/Layout';
 import DocItemTOCDesktop from '@theme/DocItem/TOC/Desktop';
 import DocItemTOCMobile from '@theme/DocItem/TOC/Mobile';
 import clsx from 'clsx';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import styles from './styles.module.css';
-import HeroHeaderByPathname from '@site/src/components/theme/HeroHeaderByPathname';
-import { PropSidebarItemLink } from '@docusaurus/plugin-content-docs';
+import ContentVisibility from '@theme/ContentVisibility';
+import DocVersionBanner from '@theme/DocVersionBanner';
 
 type Props = WrapperProps<typeof LayoutType>;
 
 export default function LayoutWrapper(props: Props): JSX.Element {
   const docType = useDocType();
-
-  if (!docType) {
-    return (
-      <div className={clsx('docLayoutWrapper', `docLayout--default`)}>
-        <Layout {...props} />
-      </div>
-    );
-  }
 
   if (docType === 'banner') {
     return <BannerDocItemLayout {...props} />;
@@ -49,8 +42,8 @@ export default function LayoutWrapper(props: Props): JSX.Element {
   }
 
   return (
-    <div className={clsx('docLayoutWrapper', `docLayout--${docType}`)}>
-      <ComponentDocItemLayout {...props} />
+    <div className={clsx('docLayoutWrapper', `docLayout--default`)}>
+      <Layout {...props} />
     </div>
   );
 }
@@ -83,7 +76,7 @@ export function BannerDocItemLayout({ children }: Props): JSX.Element {
 
   return (
     <>
-      <HeroHeader
+      <DocDefaultHeader
         id={metadata.id}
         title={title}
         description={description}
@@ -92,6 +85,8 @@ export function BannerDocItemLayout({ children }: Props): JSX.Element {
       />
       <div className={styles.Row}>
         <div className={styles.docItemContainer}>
+          <ContentVisibility metadata={metadata} />
+          <DocVersionBanner />
           <article>
             {docTOC.mobile}
 
@@ -108,62 +103,7 @@ export function BannerDocItemLayout({ children }: Props): JSX.Element {
   );
 }
 
-export function ComponentDocItemLayout({ children }: Props): JSX.Element {
-  const { currentTab, searchParams } = useSubPageHook();
-  const history = useHistory();
-  const docTOC = useDocTOC();
-  const { metadata } = useDoc();
-  const [tabs, setTabs] = useState<string[]>([]);
-  const { title, description } = metadata;
-
-  useEffect(() => {
-    if (
-      metadata.frontMatter['component-tabs'] &&
-      Array.isArray(metadata.frontMatter['component-tabs'])
-    ) {
-      setTabs(metadata.frontMatter['component-tabs']);
-    }
-  }, [metadata.frontMatter]);
-
-  useEffect(() => {
-    requestAnimationFrame(() => {
-      if (!currentTab && tabs.length > 0) {
-        searchParams.set('current-tabs', tabs[0]);
-        history.push({ search: searchParams.toString() });
-      }
-    });
-  }, [currentTab, tabs]);
-
-  return (
-    <>
-      <HeroHeader
-        id={metadata.id}
-        title={title}
-        description={description}
-        tabs={tabs}
-        frontMatter={metadata.frontMatter}
-      />
-      <div className={styles.Row}>
-        <div className={styles.docItemContainer}>
-          <article>
-            {docTOC.mobile}
-
-            <div
-              className={clsx(styles.docContent, styles.docContentComponent)}
-            >
-              <DocItemContent>{children}</DocItemContent>
-            </div>
-          </article>
-          <DocItemFooter />
-          <DocItemPaginator />
-        </div>
-        <div className={clsx('col', styles.toc)}>{docTOC.desktop}</div>
-      </div>
-    </>
-  );
-}
-
-export function DocItemTabsLayout({ children }: Props): JSX.Element {
+export function DocItemTabsLayout({}: Props): JSX.Element {
   const history = useHistory();
   const { metadata } = useDoc();
 
@@ -175,14 +115,18 @@ export function DocItemTabsLayout({ children }: Props): JSX.Element {
     }
   }, [sidebar]);
 
-  return <></>;
+  return (
+    <>
+      <ContentVisibility metadata={metadata} />
+      <DocVersionBanner />
+    </>
+  );
 }
 
 export function DocItemTabItemLayout({ children }: Props): JSX.Element {
   const docTOC = useDocTOC();
   const doc = useDoc();
   const { metadata } = doc;
-  const { title, description } = metadata;
 
   const sidebar = useCurrentSidebarCategory();
 
@@ -191,7 +135,7 @@ export function DocItemTabItemLayout({ children }: Props): JSX.Element {
 
   return (
     <>
-      <HeroHeaderByPathname
+      <DocTabsHeader
         id={metadata.id}
         title={parentDoc.title}
         description={parentDoc.description}
@@ -200,6 +144,8 @@ export function DocItemTabItemLayout({ children }: Props): JSX.Element {
       />
       <div className={styles.Row}>
         <div className={styles.docItemContainer}>
+          <ContentVisibility metadata={metadata} />
+          <DocVersionBanner />
           <article>
             {docTOC.mobile}
