+++ packages/html-test-app/src/preview-examples/html-table-striped.html
@@ -15,7 +15,7 @@ LICENSE file in the root directory of this source tree.
     <title>Html table striped example</title>
   </head>
   <body>
-    <table class="table table-striped">
+    <table class="ix-table ix-table-striped">
       <thead>
         <tr>
           <th scope="col">#</th>
