+++ packages/html-test-app/src/preview-examples/html-table.html
@@ -15,7 +15,7 @@ LICENSE file in the root directory of this source tree.
     <title>Html table example</title>
   </head>
   <body>
-    <table class="table">
+    <table class="ix-table">
       <thead>
         <tr>
           <th scope="col">#</th>
